+++ codigo/assets/js/pages/mostra_perfil.js
@@ -15,12 +15,17 @@ const btnPsi_edit = document.getElementById("btn_psi_edit");
 const btnPsi_editdesk = document.getElementById("btn_psi_edit--desk");
 const btnPsi_remove = document.getElementById("btn_psi_remove");
 const btnPsi_removedesk = document.getElementById("btn_psi_remove--desk");
+const btnUsu_edit = document.getElementById("btn_usu_edit");
+const btnUsu_editdesk = document.getElementById("btn_usu_edit--desk");
+const btnVive_remove = document.getElementById("btn_vive_remove");
 const listbtn__action = document.getElementsByClassName("list__btn--action");
 
 
 // URLs
 const URL_PSI = "http://localhost:3000/psicologos";
 const URL_ATD = "http://localhost:3000/atendimentos";
+const URL_VIVE = "http://localhost:3000/vivencias";
+const URL_USER = "http://localhost:3000/usuarios";
 let URL_PSI_API;
 
 // Variáveis Globais
@@ -89,6 +94,15 @@ function set_psi_delete(psicologo) {
    });
 }
 
+function set_vive_delete(vivencia) {
+   const id = vivencia.id;
+   const URL_VIVE_API = URL_VIVE + "/" + id;
+
+   btnVive_remove.addEventListener("click", (event) => {
+      remove_psi(URL_VIVE_API, event);
+   });
+}
+
 function listbtn_remove() {
    for (let i = 0; i < listbtn__action.length; i++) {
       listbtn__action[i].classList.remove('none');
@@ -132,7 +146,14 @@ btnFav.addEventListener("click", () => {
    toggle_page(info);
 });
 
-// Funções Assíncronas
+btnUsu_edit.addEventListener("click", (event) => {
+   console.log("Tentando editar usuario");
+});
+
+btnUsu_editdesk.addEventListener("click", (event) => {
+   console.log("Tentando editar usuario");
+});
+
 async function get_atendimentos() {
    const select = document.getElementById("select_atd");
 
@@ -187,7 +208,6 @@ async function remove_psi(url, event) {
    try {
       const response = await fetch(request);
       if (response.ok) {
-         console.log("Psicólogo removido com sucesso:", data);
          message("Psicólogo removido com sucesso", "success");
       } else {
          console.error("Erro ao remover psicólogo:", response.statusText);
@@ -199,7 +219,6 @@ async function remove_psi(url, event) {
    }
 }
 
-// Mostra os dados
 async function getFavorites_html() {
    try {
       const response_cont = await fetch("http://localhost:3000/conteudos");
@@ -213,40 +232,51 @@ async function getFavorites_html() {
 
 
       conteudos.forEach(item => {
-         const favorito = favoritos.find((fav) => fav.id_favoritos === item.id);
-         if (favorito) {
-            isFavorited = true;
-            const card = document.createElement('div');
-            card.classList.add('content-card');
-            card.innerHTML = `
-               <h3>${item.titulo}</h3>
-               <p>Categoria: ${item.categoria}</p>
-               <p>Tema: ${item.tema}</p>
-               <p>Duração: ${item.duracao} min</p>
-               <i class="fa-solid fa-bookmark favorite-icon ${isFavorited ? 'favorited' : ''}" data-id="${item.id}"></i>
-           `;
-            contentDisplay.appendChild(card);
+         let isFavorited = false;
 
-            card.querySelector('.favorite-icon').addEventListener('click', toggleFavorite);
-         }
+         favoritos.forEach(favorito => {
+            if (favorito.favorito === item.id && favorito.usuario === user.id) {
+               isFavorited = true;
+               const card = document.createElement('div');
+               card.classList.add('content-card');
+               card.innerHTML = `
+                <h3>${item.titulo}</h3>
+                <p>Categoria: ${item.categoria}</p>
+                <p>Tema: ${item.tema}</p>
+                <p>Duração: ${item.duracao} min</p>
+                <i class="fa-solid fa-bookmark favorite-icon ${isFavorited ? 'favorited' : ''}" data-id="${item.id}"></i>
+            `;
+               contentDisplay.appendChild(card);
+
+               card.querySelector('.favorite-icon').addEventListener('click', toggleFavorite);
+            }
+         });
       });
    } catch (error) {
-      console.log(error);
+      console.error(error);
    }
 }
 
-function toggleFavorite(event) {
+async function toggleFavorite(event) {
    const icon = event.target;
    const contentId = icon.getAttribute('data-id');
 
-   getFavorites().then(favorites => {
-      const favorito = favorites.find((fav) => fav.id_favoritos === contentId);
+   try {
+      const favorites = await getFavorites();
 
-      if (favorito) {
-         remove_fav(favorito.id);
-      }
-   });
+      const favorite = favorites.filter(fav => fav.favorito === contentId && fav.usuario == user.id);
+
+      favorite.forEach(async fav => {
+         await remove_fav(fav.id);
+      });
+
+      icon.classList.toggle('favorited', userFavorites.length === 0);
+   } catch (error) {
+      console.error("Erro ao alternar favorito:", error);
+      message("Erro ao alternar favorito", "error");
+   }
 }
+
 
 async function getFavorites() {
    try {
@@ -262,26 +292,37 @@ async function getFavorites() {
 
 
 // Deleta favoritos
+
 async function remove_fav(id) {
    const url = "http://localhost:3000/favoritos/" + id;
 
-   const request = new Request(url, {
-      method: "DELETE",
-      headers: { "Content-Type": "application/json" },
-   });
-
    try {
-      const response = await fetch(request);
-      if (response.ok) {
-         console.log("Favorito removido com sucesso:", data);
-         //message("Psicólogo removido com sucesso", "success");
+
+      const response = await fetch(url, {
+         method: "GET",
+         headers: { "Content-Type": "application/json" }
+      });
+      const favorite = await response.json();
+
+      if (favorite.usuario !== user.id) {
+         console.error("Este favorito não pertence ao usuário logado.");
+         return;
+      }
+
+      const deleteResponse = await fetch(url, {
+         method: "DELETE",
+         headers: { "Content-Type": "application/json" }
+      });
+
+      if (deleteResponse.ok) {
+         message("Favorito removido com sucesso", "success");
       } else {
-         console.error("Erro ao remover favorito:", response.statusText);
-         //message("Erro ao remover psicólogo", "error");
+         console.error("Erro ao remover favorito:", deleteResponse.statusText);
+         message("Erro ao remover favorito", "error");
       }
    } catch (error) {
       console.error("Erro ao remover favorito:", error);
-      //message("Erro ao remover psicólogo", "error");
+      message("Erro ao remover favorito", "error");
    }
 }
 
@@ -303,16 +344,24 @@ function fillFormPsi(form, psicologo) {
    }
 }
 
-function fillFormUser(form, user) {
-   const formData = new FormData(form);
-   formData.set("nome", user.nome);
-   formData.set("email", user.email);
+async function fillFormUser(form, id) {
+   try {
+      const usuario = await get_usuario(id);
 
-   for (const [key, value] of formData.entries()) {
-      const input = form.elements[key];
-      if (input) {
-         input.value = value;
+      const formData = new FormData(form);
+
+      formData.set("nome", usuario.nome);
+      formData.set("email", usuario.email);
+      formData.set("genero", usuario.genero);
+
+      for (const [key, value] of formData.entries()) {
+         const input = form.elements[key];
+         if (input) {
+            input.value = value;
+         }
       }
+   } catch (error) {
+      console.error("Erro ao obter usuário:", error);
    }
 }
 
@@ -324,10 +373,139 @@ function stateForm(form) {
    }
 }
 
+async function vivencias() {
+   try {
+      const response = await fetch(URL_VIVE);
+      const vivencias = await response.json();
+
+      if (response.ok) {
+         const htmlContent = await Promise.all(vivencias.map(async (vive) => {
+            try {
+               if (user.id == vive.usuario) {
+                  const usuario = await get_usuario(vive.usuario);
+
+                  const data = {
+                     id: vive.id,
+                     nome: usuario.nome,
+                     titulo: vive.titulo,
+                     situacao: vive.situacao,
+                     solucao: vive.solucao,
+                  };
+
+                  return cards_html(data);
+               }
+            } catch (error) {
+               console.error('Erro ao processar vivências:', error);
+               message("Erro ao processar as vivências", "error")
+               return '';
+            }
+         }));
+
+         publi_html.innerHTML = htmlContent.join('');
+
+         const deleteButtons = document.querySelectorAll('.btn_outline[data-id]');
+         deleteButtons.forEach(button => {
+            button.addEventListener('click', async (event) => {
+               event.preventDefault();
+               const id = button.getAttribute('data-id');
+               await delete_vivencia(id);
+            });
+         });
+      } else {
+         console.error("Erro ao buscar vivências:", response.statusText);
+         message("Erro ao buscar vivências", "error");
+      }
+   } catch (error) {
+      console.error("Erro ao buscar vivências:", error);
+      message("Erro ao buscar vivências", "error");
+   }
+}
+
+
+function cards_html(data) {
+
+   const url = '/modules/vivencias/edita_vivencias.html?id=' + data.id;
+
+   return `
+        <div class="card__vive">
+            <i class="card__quote fa-solid fa-quote-left"></i>
+            <div class="card__title">
+               <div class="card__image">
+                     <div class="card__image__icon">
+                        <i class="fa-solid fa-camera fa-2xl"></i>
+                     </div>
+               </div>
+
+               <div class="card__title--info">
+                     <h4 class="card__name">${data.titulo}</h4>
+                     <p class="card__type">${data.nome}</p>
+               </div>
+            </div>
+
+            <div class="card__info">
+               <ul class="card__info__lista">
+                     <li class="card__info__item">
+                        <h6><b>Situação:</b></h6>
+                        <p>${data.situacao}</p>
+                     </li>
+                     <li class="card__info__item">
+                        <h6><b>Solução:</b></h6>
+                        <p>${data.solucao}</p>
+                     </li>
+               </ul>
+            </div>
+
+            <div class="card__footer">
+               <a href="${url}" class="btn_outline"> <i class="fa-solid fa-pencil"></i>Editar</a>
+               <a href="#" class="btn_outline" data-id="${data.id}"> <i class="fa-solid fa-trash" id="btn_vive_remove"></i>Excluir</a>
+            </div>
+         </div>
+       `;
+}
+
+
+async function get_usuario(id) {
+   try {
+      const response = await fetch(URL_USER);
+      const usuarios = await response.json();
+
+      const usuario = usuarios.find((u) => u.id === id);
+
+      if (usuario) {
+         return usuario;
+      }
+   } catch (error) {
+      return console.error("Erro ao buscar usuario:", error);
+   }
+}
+
+async function delete_vivencia(id) {
+   const url = URL_VIVE + "/" + id;
+
+   try {
+      const response = await fetch(url, {
+         method: "DELETE",
+         headers: { "Content-Type": "application/json" }
+      });
+
+      if (response.ok) {
+         message("Vivência removida com sucesso", "success");
+         await vivencias();
+      } else {
+         console.error("Erro ao remover vivência:", response.statusText);
+         message("Erro ao remover vivência", "error");
+      }
+   } catch (error) {
+      console.error("Erro ao remover vivência:", error);
+      message("Erro ao remover vivência", "error");
+   }
+}
+
 // Inicialização
 check();
 fetchPsicologo();
 toggle_page(info);
-fillFormUser(formUser, user);
+fillFormUser(formUser, user.id);
 get_atendimentos();
 getFavorites_html();
+vivencias();
